refactor: fileSize 타입 변경
refactor: fileSize 타입 변경

- fileSize 타입 Double -> Long 으로 변경

+++ media/src/main/kotlin/com/fx/media/domain/Media.kt
@@ -7,7 +7,7 @@ data class Media (
     val id: Long?= null,
     val fileUrl: String,
     val fileType: FileType,
-    val fileSize: Double,
+    val fileSize: Long,
     val userId: Long?= null,
     val serverName: String,
     val originalName: String,
